liang

+++ src/main/java/com/yangnan/selfhelpordingsystem/controller/CookInfoController.java
@@ -3,6 +3,7 @@ package com.yangnan.selfhelpordingsystem.controller;
 import com.yangnan.selfhelpordingsystem.common.CommonResult;
 import com.yangnan.selfhelpordingsystem.constant.BillDetailStatus;
 import com.yangnan.selfhelpordingsystem.constant.CookStatus;
+import com.yangnan.selfhelpordingsystem.constant.SessionParameters;
 import com.yangnan.selfhelpordingsystem.dto.BillDetailDTO;
 import com.yangnan.selfhelpordingsystem.dto.CookDTO;
 import com.yangnan.selfhelpordingsystem.dto.GoodsDTO;
@@ -18,6 +19,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -44,7 +47,6 @@ public class CookInfoController {
                                        @RequestParam(required = false, defaultValue = "") String telephone,
                                        @RequestParam(required = false, defaultValue = "") Integer status) {
         List<CookDTO> cookDTOList = cookService.queryCookInfo(name, telephone, status);
-        System.out.println(cookDTOList);
         if (CollectionUtils.isEmpty(cookDTOList)) {
             return CommonResult.fail(404, "暂无相关信息");
         }
@@ -54,17 +56,19 @@ public class CookInfoController {
     /**
      * 动态修改厨师信息
      *
+     * @param request
      * @param cookName
      * @param telephone
-     * @param cookId
      * @return
      */
-    @GetMapping("/update/cookInfo")
-    public CommonResult updateCookInfo(Integer cookId,
+    @GetMapping("/cook/update/cookInfo")
+    public CommonResult updateCookInfo(HttpServletRequest request,
                                        @RequestParam(required = false, defaultValue = "") String cookName,
                                        @RequestParam(required = false, defaultValue = "") String telephone,
                                        @RequestParam(required = false, defaultValue = "") String nickname,
                                        @RequestParam(required = false, defaultValue = "") String cookPassword) {
+        HttpSession session = request.getSession();
+        Integer cookId = (Integer) session.getAttribute(SessionParameters.COOKID);
         if (cookId == null) {
             return CommonResult.fail(403, "参数错误");
         }
@@ -116,10 +120,13 @@ public class CookInfoController {
      * @param cookId
      * @return
      */
-    @GetMapping("/cook/delete/info")
+    @GetMapping("/manager/delete/cookInfo")
     public CommonResult deleteCookInfo(Integer cookId) {
         if (cookId == null) {
             return CommonResult.fail(403, "参数错误");
+        }
+        if (selectCookOrder(cookId)) {
+            return CommonResult.fail(403, "暂时无法删除厨师，正在做菜");
         }
         int result = cookService.deleteCookerInfo(cookId);
         if (result <= 0) {
@@ -129,7 +136,7 @@ public class CookInfoController {
     }
 
     @GetMapping("/cook/login")
-    public CommonResult login(String username, String password) {
+    public CommonResult login(String username, String password, HttpServletRequest request) {
         if (Strings.isEmpty(username) || Strings.isEmpty(password)) {
             return CommonResult.fail(403, "参数错误");
         }
@@ -137,24 +144,50 @@ public class CookInfoController {
         if (cookDTO == null) {
             return CommonResult.fail(404, "查不到对应用户");
         }
-        cookService.updateStatusById(cookDTO.getId(), CookStatus.WORK);
-        return CommonResult.success();
+        try {
+            HttpSession session = request.getSession();
+            session.setAttribute(SessionParameters.USERNAME, username);
+            session.setAttribute(SessionParameters.PASSWORD, password);
+            session.setAttribute(SessionParameters.COOKID, cookDTO.getId());
+            cookService.updateStatusById(cookDTO.getId(), CookStatus.WORK);
+            return CommonResult.success();
+        } catch (Exception e) {
+            System.out.println("登录失败");
+        }
+        return CommonResult.fail(500, "服务器异常");
     }
 
     @GetMapping("/cook/loginOut")
-    public CommonResult loginOut(Integer cookId) {
+    public CommonResult loginOut(HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        Integer cookId = (Integer) session.getAttribute(SessionParameters.COOKID);
         if (cookId == null || cookId < 1) {
             return CommonResult.fail(403, "参数错误");
         }
+        if (selectCookOrder(cookId)) {
+            return CommonResult.fail(403, "参数错误,无法休息，有正在做的菜");
+        }
+        try {
+            session.setAttribute(SessionParameters.USERNAME, "");
+            session.setAttribute(SessionParameters.PASSWORD, "");
+            session.setAttribute(SessionParameters.COOKID, 0);
+            cookService.updateStatusById(cookId, CookStatus.REST);
+            return CommonResult.success(200, "退出登录成功");
+        } catch (Exception e) {
+            System.out.println("退出登录失败");
+        }
+        return CommonResult.fail(500, "服务器异常");
+    }
+
+    private Boolean selectCookOrder(Integer cookId) {
         List<GoodsDTO> goodsDTOS = goodsService.selectGoodsByCookId(cookId);
         if (!CollectionUtils.isEmpty(goodsDTOS)) {
             List<Integer> goodsIds = goodsDTOS.stream().map(GoodsDTO::getId).collect(Collectors.toList());
             List<BillDetailDTO> billDetailDTOS = billDetailService.selectOrderByGoodsIds(goodsIds, BillDetailStatus.PRODUCING);
             if (!CollectionUtils.isEmpty(billDetailDTOS)) {
-                return CommonResult.fail(403, "你还有正在做的菜，无法休息，参数错误");
+                return true;
             }
         }
-        cookService.updateStatusById(cookId, CookStatus.REST);
-        return CommonResult.success();
+        return false;
     }
 }
